refactor: SortAlert 메서드 확장으로 재정의
